requirements.txt
Django==4.1.1
djangorestframework==3.13.1
djangorestframework_simplejwt==5.2.0
requests==2.28.1
